[MINSTALL-160] Generated POM is not installed if original POM exists
When using `generatePom=true` with install-file Mojo, it will despite
this copy existing POM if it is found embedded in JAR (and if it
was built with Maven, it will find it).

The generatePom should reuse POM values embedded from JAR
but should generate (and install) new "empty POM" instead.